Corrections: map, cmake, filters bounding area
Corrections
- in main CMakeLists.txt to compile externals
- in env-co.pt to copy python32.dll dependency
- in map
- in algorithms dialog

Implementations:
- proper (fast) implementation on BoundingArea (however, may get **SLOW** if CArea::removeVertex() is called many times).
- handleAreasSelectionChanged() implemented according to our previous CArea min/max implementation.

+++ new-gui/brat/GUI/DisplayWidgets/MapWidget.cpp
@@ -1111,7 +1111,7 @@ QgsGraduatedSymbolRendererV2* CMapWidget::CreateRenderer( const QString &target_
     auto range_list = new QgsRangeList;
 
 	const double range = M - m;
-    size_t step = range / contours;	// color_map_size;
+    double step = range / contours;	// color_map_size;
 
 	mc.ResetPrimaryColors();
 	double v = m;
